Сhanging int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{5, 16, 7, 28, 0, 55});
+        SalesManager salesManager = new SalesManager(new long []{5, 16, 7, 28, 0, 55});
 
         System.out.println("Максимальное значение = " + salesManager.max());
     }
